Renamed DevNamesCommand to CreditsCommand, updated description, and changed formatting of embed.
